// src/app/models/rep_stockrealRPT.ts
export class RepStockRealRPT {
    Id: number;
    Container: string;
    Size: number;
    Tipo: string;
    Tara: number;
    PayLoad: number;
    AnioFabr: number;
    Clase: string;
    CondicionCaja: string;
    FechaRecepcion: string;
    Nave: string;
    NroViaje: string;
  }

  export class RepStockRealRefRPT {

    Contenedor: string;
    Size: number;
    Tipo: number;
    Tara: number;
    PayLoad: number;
    AñoFabr: number;
    Clase: string;
    CondicioónCaja: string;
    CondicioónMaquina: string
    Bloqueo: string;
    BloqueoMotivo: string;
    FRecepcion: string;
    Nave: string;
    NroViaje: string;
    Marca: string;

  }

  
  /*Id: number;
  Container: string;
  Size: number;
  Tipo: string;
  Tara: number;
  PayLoad: number;
  AnioFabr: number;
  Clase: string;
  CondicionCaja: string;
  CondicionMaquina: string;
  Bloqueo: string;
  BloqueoMotivo: string;
  FechaRecepcion: string;
  Nave: string;
  NroViaje: string;
  Marca: string;*/

  
  export class RepStockRealCabecera {
    Id: number;
    Columna: string;
  }